refactor(CircuitBreakerDecorator): moving decorator in root folder
