+++ src/index.ts
@@ -1,4 +1,4 @@
 export { CircuitBreaker } from './CircuitBreaker';
 export { CircuitBreakerState } from './CircuitBreakerState';
 export { CircuitBreakerError } from './CircuitBreakerError';
-export { CircuitBreakerDecorator } from './decorator/CircuitBreakerDecorator';
+export { CircuitBreakerDecorator } from './CircuitBreakerDecorator';
